airflow/requirements.txt
apache-airflow[gcp]==2.9.1
python-dotenv>=1.0.0